requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
alembic==1.16.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
appdirs==1.4.4
asteroid-filterbanks==0.4.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
av==15.0.0
backoff==2.2.1
bcrypt==4.3.0
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
chromadb==1.0.15
click==8.2.1
coloredlogs==15.0.1
colorlog==6.9.0
contourpy==1.3.2
cryptography==45.0.5
ctranslate2==4.4.0
cycler==0.12.1
decorator==5.2.1
diskcache==5.6.3
distro==1.9.0
docopt==0.6.2
docstring_parser==0.17.0
durationpy==0.10
einops==0.8.1
et_xmlfile==2.0.0
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.116.1
faster-whisper==1.1.0
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2025.7.0
google-auth==2.40.3
googleapis-common-protos==1.70.0
grpcio==1.74.0
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.34.1
humanfriendly==10.0
HyperPyYAML==1.2.2
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
instructor==1.10.0
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
json5==0.12.0
json_repair==0.25.2
jsonpickle==4.1.1
jsonref==1.1.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
julius==0.2.7
kiwisolver==1.4.8
kubernetes==33.1.0
lightning==2.5.2
lightning-utilities==0.15.0
litellm==1.74.3
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mmh3==5.1.0
mpmath==1.3.0
multidict==6.6.3
networkx==3.4.2
nltk==3.9.1
numpy==2.2.6
oauthlib==3.3.1
omegaconf==2.3.0
onnxruntime==1.22.0
openai==1.97.1
openpyxl==3.1.5
opentelemetry-api==1.35.0
opentelemetry-exporter-otlp-proto-common==1.35.0
opentelemetry-exporter-otlp-proto-grpc==1.35.0
opentelemetry-exporter-otlp-proto-http==1.35.0
opentelemetry-proto==1.35.0
opentelemetry-sdk==1.35.0
opentelemetry-semantic-conventions==0.56b0
optuna==4.4.0
orjson==3.11.1
overrides==7.7.0
packaging==25.0
pandas==2.3.1
parso==0.8.4
pdfminer.six==20250506
pdfplumber==0.11.7
pexpect==4.9.0
pillow==11.3.0
portalocker==2.7.0
posthog==5.4.0
primePy==1.3
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyannote.audio==3.3.2
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.3
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytorch-lightning==2.5.2
pytorch-metric-learning==2.8.1
pytz==2025.2
pyvis==0.3.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
rich==14.1.0
rpds-py==0.26.0
rsa==4.9.1
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.15.3
semver==3.0.4
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
speechbrain==1.0.3
SQLAlchemy==2.0.41
stack-data==0.6.3
starlette==0.47.2
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
tensorboardX==2.6.4
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.2
tomli==2.2.1
tomli_w==1.2.0
torch==2.2.2
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchaudio==2.2.2
torchmetrics==1.8.0
tqdm==4.67.1
traitlets==5.14.3
transformers==4.54.0
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uv==0.8.3
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
whisperx==3.3.1
yarl==1.20.1
zipp==3.23.0
